% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faIcons.R
\name{addFaLegend}
\alias{addFaLegend}
\title{Add a Font Awesome legend to a map}
\usage{
addFaLegend(
  map,
  icons,
  labels,
  colors = "black",
  title = NULL,
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  layerId = NULL,
  className = "info legend",
  data = leaflet::getMapData(map)
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{icons}{Name of the Font Awesome icons, passed to \code{\link[fontawesome:fa]{fontawesome::fa()}}. A
full list of available icons can be found using
\code{\link[fontawesome:fa_metadata]{fontawesome::fa_metadata()}}.}

\item{labels}{Labels for each \code{icon}; should be the same length as \code{icons}.}

\item{colors}{Colours to use for each \code{icon}. If \code{length(colors) == 1L} it is
recycled for all icons. Otherwise should be the same length as \code{icons}.}

\item{title}{the legend title; optional.}

\item{position}{position of control: "topleft", "topright", "bottomleft", or
"bottomright"}

\item{layerId}{the layer id}

\item{className}{extra CSS classes to append to the control, space separated}

\item{data}{the data object from which the argument values are derived; by
default, it is the \code{data} object provided to \code{leaflet()}
initially, but can be overridden}
}
\value{
a \link[leaflet:leaflet-package]{leaflet} widget
}
\description{
Manually specify icons, labels, and colours to add a Font Awesome legend to a
map. Useful in conjunction with \code{\link[=faIcons]{faIcons()}} to communicate the meaning of
icons/colours.
}
\examples{
library(leaflet)
addFaLegend(
  map = leaflet(),
  icons = c("beer", "school"),
  labels = c("Pub", "University"),
  colors = c("red", "blue"),
  title = "Student Hangout"
)
}
