% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faIcons.R
\name{faIcons}
\alias{faIcons}
\title{Create a set of Font Awesome Markers}
\usage{
faIcons(
  icon = "circle",
  markerColor = NULL,
  iconColor = NULL,
  className = NULL
)
}
\arguments{
\item{icon}{Name of the Font Awesome icon, passed to \code{\link[fontawesome:fa]{fontawesome::fa()}}. A
full list of available icons can be found using
\code{\link[fontawesome:fa_metadata]{fontawesome::fa_metadata()}}.}

\item{markerColor}{The color of the teardrop-shaped marker.}

\item{iconColor}{The color of the fontawesome icon.}

\item{className}{a custom class name to assign to both icon and shadow images}
}
\value{
a \code{\link[leaflet:iconList]{leaflet::iconList()}}, to be passed to the \code{icon} argument of
\code{\link[leaflet:map-layers]{leaflet::addMarkers()}}
}
\description{
This function is intended as a more up-to-date implementation of
\code{\link[leaflet:awesomeIcons]{leaflet::awesomeIcons()}}. Key benefits include having access to modern Font
Awesome icons, and allowing any colour to be used the marker and icon.
}
\examples{
# create map
library(leaflet)
port_talbot \%>\%
  dplyr::mutate(
    icon = dplyr::case_match(
      site_type,
      "Urban Industrial" ~ "industry",
      "Urban Traffic" ~ "car",
      "Urban Background" ~ "city",
      .default = "x"
    ),
    color = dplyr::case_match(
      site_type,
      "Urban Industrial" ~ "#12436D",
      "Urban Traffic" ~ "#801650",
      "Urban Background" ~ "#28A197",
      .default = "#3D3D3DFF"
    )
  ) \%>\%
  leaflet() \%>\%
  addTiles() \%>\%
  addMarkers(icon = ~ faIcons(icon, color, "white"),
             popup = ~ site)
}
\author{
Jack Davison
}
