% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.magic-package.R
\docType{package}
\name{leaf.magic-package}
\alias{leaf.magic}
\alias{leaf.magic-package}
\title{leaf.magic: Customisable Icon Markers for 'leaflet'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use modern Icon libraries to construct customisable leaflet marker icons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-davison.github.io/leaf.magic/}
  \item \url{https://github.com/jack-davison/leaf.magic}
  \item Report bugs at \url{https://github.com/jack-davison/leaf.magic/issues}
}

}
\author{
\strong{Maintainer}: Jack Davison \email{davison.jack.jd@gmail.com} (\href{https://orcid.org/0000-0003-2653-6615}{ORCID})

}
\keyword{internal}
