% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicIcons.R
\name{magicIcons}
\alias{magicIcons}
\title{Create a set of Font Awesome, Bootstrap or Ionicons Markers}
\usage{
magicIcons(
  icon = "circle",
  markerColor = leaf.magic::awesomePalette$blue,
  iconColor = leaf.magic::awesomePalette$white,
  marker = "marker",
  markerSize = 30L,
  library = "fontawesome",
  className = NULL,
  dir = tempdir()
)
}
\arguments{
\item{icon}{Name of the icon. A full list of available icons can be found
using \code{\link[fontawesome:fa_metadata]{fontawesome::fa_metadata()}} or the equivalent function for other
\code{library} options.}

\item{markerColor}{The color of the marker. Not used when \code{marker = "none"}.}

\item{iconColor}{The color of the icon.}

\item{marker}{Defaults to \code{"marker"}, which uses the standard teardrop shaped
marker, similar to \code{\link[leaflet:map-layers]{leaflet::addMarkers()}}. Other options are \code{"circle"},
\code{"square"}, \code{"star"}, \code{"heart"}, and \code{"diamond"}, which place the icon
inside of the respective shape. Also available is \code{"none"}, which removes
the marker entirely and places the icon directly on the map.}

\item{markerSize}{The size of the marker. Defaults to \code{30}, which is roughly
the same size as \code{\link[leaflet:map-layers]{leaflet::addMarkers()}}.}

\item{library}{One of \code{"fontawesome"}, \code{"bootstrap"}, \code{"ionicons"}, or
\code{"lucide"} defining the icon library of interest. Defaults to
\code{"fontawesome"}.}

\item{className}{a custom class name to assign to both icon and shadow images}

\item{dir}{The directory in which markers are saved. By default this is
\code{\link[=tempdir]{tempdir()}}, which is a temporary directory after each session. Providing
an alternative directory will allow markers to persist between R sessions.}
}
\value{
a \code{\link[leaflet:iconList]{leaflet::iconList()}}, to be passed to the \code{icon} argument of
\code{\link[leaflet:map-layers]{leaflet::addMarkers()}}
}
\description{
This function is intended as a more up-to-date implementation of
\code{\link[leaflet:awesomeIcons]{leaflet::awesomeIcons()}}. Key benefits include having access to modern Font
Awesome, Bootstrap and Ionicons icons, and allowing any colour to be used the
marker and icon.
}
\details{
This function uses \code{\link[magick:composite]{magick::image_composite()}} to knit together the icon and
marker symbol of choice, and saves it to a temporary directory. It is also
intelligent enough to not try to recreate a marker that already exists, so
you may find the use of \code{\link[=magicIcons]{magicIcons()}} speeds up the more you use it in a
single session.
}
\examples{
# create map
library(leaflet)
port_talbot |>
  dplyr::mutate(
    icon = dplyr::recode_values(
      site_type,
      "Urban Industrial" ~ "industry",
      "Urban Traffic" ~ "car",
      "Urban Background" ~ "city",
      default = "x"
    ),
    color = dplyr::recode_values(
      site_type,
      "Urban Industrial" ~ "#12436D",
      "Urban Traffic" ~ "#801650",
      "Urban Background" ~ "#28A197",
      default = "#3D3D3DFF"
    )
  ) |>
  leaflet() |>
  addProviderTiles("CartoDB.Voyager") |>
  addMarkers(
    icon = ~ magicIcons(icon, color, "white"),
    lat = ~ latitude,
    lng = ~ longitude,
    popup = ~site
  )
}
\seealso{
\code{\link[=addIconLegend]{addIconLegend()}}, \code{\link[=iconFactor]{iconFactor()}}, \code{\link[=iconBin]{iconBin()}}, \code{\link[=iconQuantile]{iconQuantile()}}
}
\author{
Jack Davison
}
