% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_talbot.R
\docType{data}
\name{port_talbot}
\alias{port_talbot}
\title{Air Quality Monitoring in Neath Port Talbot, 2024}
\format{
\subsection{\code{port_talbot}}{

A data frame with 9 rows and 8 columns:
\describe{
\item{code}{The unique code for the measurement site}
\item{site}{The human-readable site name for each site}
\item{latitude,longitude}{The decimal latitude/longitude coordinates of the site}
\item{site_type}{The site classification of the site; in this case, either Urban Industrial, Traffic, or Background. See \url{https://www.airquality.gov.wales/about-air-quality/monitoring/monitoring-locations} for more information.}
\item{open}{Logical; was the site still open at time of data compilation?}
\item{start_date,end_date}{The opening and closing date of the site. An \code{NA} in \code{end_date} means the site is still open}
}
}
}
\source{
\url{https://www.airquality.gov.wales/}
}
\usage{
port_talbot
}
\description{
A subset of fixed monitoring stations which make up the Welsh Air Quality
Network in Neath Port Talbot, Wales. Data were obtained and correct as of
July 2024. This dataset is provided for demonstration of \code{leaf.magick}
functions.
}
\examples{
# view data
port_talbot
}
\keyword{datasets}
