% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{addIconAttribution}
\alias{addIconAttribution}
\title{Conveniently add Icon Attributions to a Leaflet Map}
\usage{
addIconAttribution(map, library = "fontawesome", layerId = NULL, group = NULL)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{library}{One of \code{"fontawesome"}, \code{"bootstrap"}, or \code{"ionicons"},
defining the icon library of interest. Defaults to \code{"fontawesome"}.}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for \code{\link[leaflet]{clearGroup}} and \code{\link[leaflet]{addLayersControl}} purposes).
Human-friendly group names are permitted--they need not be short,
identifier-style names. Any number of layers and even different types of
layers (e.g. markers and polygons) can share the same group name.}
}
\description{
Adds a short character string to the bottom-right of the map widget with the
attribution and license of an icon library. \code{group} and \code{layerId} can be used
to add or remove the attribution - e.g., if the markers using \code{\link[=magicIcons]{magicIcons()}}
are hidden.
}
