% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicIcons.R
\name{addIconLegend}
\alias{addIconLegend}
\title{Add an Icon legend to a map}
\usage{
addIconLegend(
  map,
  icons,
  labels,
  colors = "black",
  title = NULL,
  library = "fontawesome",
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  layerId = NULL,
  className = "info legend",
  data = leaflet::getMapData(map)
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{icons}{Name of the icons, passed to \code{\link[fontawesome:fa]{fontawesome::fa()}} or \code{\link[bsicons:bs_icon]{bsicons::bs_icon()}}. A
full list of available icons can be found using
\code{\link[fontawesome:fa_metadata]{fontawesome::fa_metadata()}} or at \url{https://icons.getbootstrap.com/}.}

\item{labels}{Labels for each \code{icon}; should be the same length as \code{icons}.}

\item{colors}{Colours to use for each \code{icon}. If \code{length(colors) == 1L} it is
recycled for all icons. Otherwise should be the same length as \code{icons}.}

\item{title}{the legend title; optional.}

\item{library}{One of \code{"fontawesome"} or \code{"bootstrap"}, defining the icon
library of interest. Defaults to \code{"fontawesome"}.}

\item{position}{position of control: "topleft", "topright", "bottomleft", or
"bottomright"}

\item{layerId}{the layer id}

\item{className}{extra CSS classes to append to the control, space separated}

\item{data}{the data object from which the argument values are derived; by
default, it is the \code{data} object provided to \code{leaflet()}
initially, but can be overridden}
}
\value{
a \link[leaflet:leaflet-package]{leaflet} widget
}
\description{
Manually specify icons, labels, and colours to add an Icon legend to a
map. Useful in conjunction with \code{\link[=magicIcons]{magicIcons()}} to communicate the meaning of
icons/colours.
}
\examples{
library(leaflet)
addIconLegend(
  map = leaflet(),
  icons = c("beer", "school"),
  labels = c("Pub", "University"),
  colors = c("red", "blue"),
  title = "Student Hangout"
)
}
