% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.magick-package.R
\docType{package}
\name{leaf.magick-package}
\alias{leaf.magick}
\alias{leaf.magick-package}
\title{leaf.magick: Customisable Icon Markers for 'leaflet'}
\description{
Use the up-to-date Font Awesome library to construct customisable leaflet marker icons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-davison.github.io/leaf.magick/}
}

}
\author{
\strong{Maintainer}: Jack Davison \email{davison.jack.jd@gmail.com} (\href{https://orcid.org/0000-0003-2653-6615}{ORCID})

}
\keyword{internal}
