% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicIcons.R
\name{magicIcons}
\alias{magicIcons}
\title{Create a set of Font Awesome, Bootstrap or Ionicons Markers}
\usage{
magicIcons(
  icon = "circle",
  markerColor = awesomePalette$blue,
  iconColor = awesomePalette$white,
  markerSize = 30L,
  library = "fontawesome",
  className = NULL
)
}
\arguments{
\item{icon}{Name of the Font Awesome icon, passed to \code{\link[fontawesome:fa]{fontawesome::fa()}} or
\code{\link[bsicons:bs_icon]{bsicons::bs_icon()}}. A full list of available icons can be found using
\code{\link[fontawesome:fa_metadata]{fontawesome::fa_metadata()}} or at \url{https://icons.getbootstrap.com/}.}

\item{markerColor}{The color of the teardrop-shaped marker.}

\item{iconColor}{The color of the fontawesome icon.}

\item{markerSize}{The size of the marker. Defaults to \code{30}, which is roughly
the same size as \code{\link[leaflet:map-layers]{leaflet::addMarkers()}}.}

\item{library}{One of \code{"fontawesome"}, \code{"bootstrap"}, or \code{"ionicons"},
defining the icon library of interest. Defaults to \code{"fontawesome"}.}

\item{className}{a custom class name to assign to both icon and shadow images}
}
\value{
a \code{\link[leaflet:iconList]{leaflet::iconList()}}, to be passed to the \code{icon} argument of
\code{\link[leaflet:map-layers]{leaflet::addMarkers()}}
}
\description{
This function is intended as a more up-to-date implementation of
\code{\link[leaflet:awesomeIcons]{leaflet::awesomeIcons()}}. Key benefits include having access to modern Font
Awesome, Bootstrap and Ionicons icons, and allowing any colour to be used the
marker and icon.
}
\details{
This function uses \code{\link[magick:composite]{magick::image_composite()}} to knit together the icon and
marker symbol of choice, and saves it to a temporary directory. It is also
intelligent enough to not try to recreate a marker that already exists, so
you may find the use of \code{\link[=magicIcons]{magicIcons()}} speeds up the more you use it in a
single session.
}
\examples{
# create map
library(leaflet)
port_talbot \%>\%
  dplyr::mutate(
    icon = dplyr::case_match(
      site_type,
      "Urban Industrial" ~ "industry",
      "Urban Traffic" ~ "car",
      "Urban Background" ~ "city",
      .default = "x"
    ),
    color = dplyr::case_match(
      site_type,
      "Urban Industrial" ~ "#12436D",
      "Urban Traffic" ~ "#801650",
      "Urban Background" ~ "#28A197",
      .default = "#3D3D3DFF"
    )
  ) \%>\%
  leaflet() \%>\%
  addProviderTiles("CartoDB.Voyager") \%>\%
  addMarkers(
    icon = ~ magicIcons(icon, color, "white"),
    popup = ~site
  )
}
\seealso{
\code{\link[=addIconLegend]{addIconLegend()}}, \code{\link[=iconFactor]{iconFactor()}}, \code{\link[=iconBin]{iconBin()}}, \code{\link[=iconQuantile]{iconQuantile()}}
}
\author{
Jack Davison
}
